+++ src/main/java/msg/team1/Hi/domain/home_base/service/HomeBaseService.java
@@ -1,13 +1,8 @@
 package msg.team1.Hi.domain.home_base.service;
 
 import msg.team1.Hi.domain.home_base.presentation.dto.request.ReserveHomeBaseRequest;
-import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationDetailResponse;
-import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationResponse;
-
-import java.util.List;
 
 public interface HomeBaseService {
     void reserveHomeBase(ReserveHomeBaseRequest request);
-    List<LookUpReservationResponse> lookUpAllReservation(Integer floor, Integer period);
-    LookUpReservationDetailResponse lookUpReservation(Long reservationId);
+
 }
