+++ src/main/java/msg/team1/Hi/domain/home_base/service/impl/HomeBaseServiceImpl.java
@@ -6,16 +6,12 @@ import msg.team1.Hi.domain.home_base.exception.ForbiddenHomeBaseReservationExcep
 import msg.team1.Hi.domain.home_base.exception.FullHomeBaseReservationException;
 import msg.team1.Hi.domain.home_base.exception.NotFoundHomeBaseException;
 import msg.team1.Hi.domain.home_base.presentation.dto.request.ReserveHomeBaseRequest;
-import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationDetailResponse;
-import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationResponse;
 import msg.team1.Hi.domain.home_base.repository.HomeBaseRepository;
 import msg.team1.Hi.domain.home_base.service.HomeBaseService;
 import msg.team1.Hi.domain.member.entity.Member;
 import msg.team1.Hi.domain.member.entity.enum_type.UseStatus;
-import msg.team1.Hi.domain.member.presentation.dto.response.MemberInfoResponse;
 import msg.team1.Hi.domain.reservation.entity.Reservation;
 import msg.team1.Hi.domain.reservation.entity.enum_type.CheckStatus;
-import msg.team1.Hi.domain.reservation.exception.NotFoundReservationException;
 import msg.team1.Hi.domain.reservation.repository.ReservationRepository;
 import msg.team1.Hi.global.annotation.TransactionalService;
 import msg.team1.Hi.global.util.HomeBaseUtil;
@@ -30,7 +26,6 @@ public class HomeBaseServiceImpl implements HomeBaseService {
     private final MemberUtil memberUtil;
     private final HomeBaseRepository homeBaseRepository;
     private final ReservationRepository reservationRepository;
-    private final HomeBaseUtil homeBaseUtil;
 
     private void isAvailableHomeBaseAndMember(HomeBase homeBase, Member member) {
         if(homeBase.isFull())
@@ -72,28 +67,5 @@ public class HomeBaseServiceImpl implements HomeBaseService {
         reservationRepository.save(reservation);
     }
 
-    @Override
-    public List<LookUpReservationResponse> lookUpAllReservation(Integer floor, Integer period) {
-        HomeBase homeBase = homeBaseRepository.findByFloorAndPeriod(floor, period)
-                .orElseThrow(() -> new NotFoundHomeBaseException("존재하지 않는 홈베이스입니다."));
-
-        List<Reservation> reservations = reservationRepository.findAllByHomeBase(homeBase);
-
-        return homeBaseUtil.reservationToLookUpResponseDto(reservations);
-    }
-
-    @Override
-    public LookUpReservationDetailResponse lookUpReservation(Long reservationId) {
-        Reservation reservation = reservationRepository.findById(reservationId)
-                .orElseThrow(() -> new NotFoundReservationException("존재하지 않는 예약현황입니다."));
-
-        List<MemberInfoResponse> memberInfoList = memberUtil.memberListToDtoList(reservation.getMembers());
-
-        return LookUpReservationDetailResponse.builder()
-                .teamName(reservation.getTeamName())
-                .representative(memberUtil.memberToDto(reservation.getRepresentative()))
-                .members(memberInfoList)
-                .build();
-    }
 
 }
