+++ src/main/java/msg/team1/Hi/domain/reservation/service/impl/ReservationServiceImpl.java
@@ -1,0 +1,53 @@
+package msg.team1.Hi.domain.reservation.service.impl;
+
+import lombok.RequiredArgsConstructor;
+import msg.team1.Hi.domain.home_base.entity.HomeBase;
+import msg.team1.Hi.domain.home_base.exception.NotFoundHomeBaseException;
+import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationDetailResponse;
+import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationResponse;
+import msg.team1.Hi.domain.home_base.repository.HomeBaseRepository;
+import msg.team1.Hi.domain.member.presentation.dto.response.MemberInfoResponse;
+import msg.team1.Hi.domain.reservation.entity.Reservation;
+import msg.team1.Hi.domain.reservation.exception.NotFoundReservationException;
+import msg.team1.Hi.domain.reservation.repository.ReservationRepository;
+import msg.team1.Hi.domain.reservation.service.ReservationService;
+import msg.team1.Hi.global.annotation.TransactionalService;
+import msg.team1.Hi.global.util.HomeBaseUtil;
+import msg.team1.Hi.global.util.MemberUtil;
+
+import java.util.List;
+
+@TransactionalService
+@RequiredArgsConstructor
+public class ReservationServiceImpl implements ReservationService {
+
+    private final MemberUtil memberUtil;
+    private final HomeBaseRepository homeBaseRepository;
+    private final ReservationRepository reservationRepository;
+    private final HomeBaseUtil homeBaseUtil;
+
+    @Override
+    public List<LookUpReservationResponse> lookUpAllReservation(Integer floor, Integer period) {
+        HomeBase homeBase = homeBaseRepository.findByFloorAndPeriod(floor, period)
+                .orElseThrow(() -> new NotFoundHomeBaseException("존재하지 않는 홈베이스입니다."));
+
+        List<Reservation> reservations = reservationRepository.findAllByHomeBase(homeBase);
+
+        return homeBaseUtil.reservationToLookUpResponseDto(reservations);
+    }
+
+    @Override
+    public LookUpReservationDetailResponse lookUpReservation(Long reservationId) {
+        Reservation reservation = reservationRepository.findById(reservationId)
+                .orElseThrow(() -> new NotFoundReservationException("존재하지 않는 예약현황입니다."));
+
+        List<MemberInfoResponse> memberInfoList = memberUtil.memberListToDtoList(reservation.getMembers());
+
+        return LookUpReservationDetailResponse.builder()
+                .teamName(reservation.getTeamName())
+                .representative(memberUtil.memberToDto(reservation.getRepresentative()))
+                .members(memberInfoList)
+                .build();
+    }
+
+}
